requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 音声処理
openai-whisper==20231117
opensmile==2.5.0
librosa==0.10.1
soundfile==0.12.1
noisereduce==3.0.0
pydub==0.25.1

# 機械学習
transformers==4.35.2
torch==2.1.1
xgboost==2.0.2
scikit-learn==1.3.2

# データ処理
numpy==1.26.2
pandas==2.1.3

# その他
python-dotenv==1.0.0
aiofiles==23.2.1